requirements_수정.txt
absl-py==2.1.0
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
astunparse==1.6.3
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.8
debugpy==1.8.1
decorator==5.1.1
dnspython==2.5.0
docker-pycreds==0.4.0
easydict==1.11
exceptiongroup==1.2.1
executing==2.0.1
expecttest==0.2.1
flatbuffers==24.3.25
fonttools==4.48.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.63.0
huggingface-hub==0.23.0
humanfriendly==10.0
hydra-core==1.3.2
hypothesis==6.97.3
imageio==2.34.0
insightface==0.7.3
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
joblib==1.3.2
jsonpointer==2.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.3
Markdown==3.6
matplotlib==3.8.2
matplotlib-inline==0.1.7
mkl-service==2.4.0
more-itertools==9.2.0
nest-asyncio==1.6.0
onnx==1.15.0
onnxruntime==1.18.0
opencv-python-headless==4.9.0.80
optree==0.10.0
packaging==24.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prettytable==3.9.0
prompt_toolkit==3.0.44
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-etcd==0.4.5
pyzmq==26.0.3
qudida==0.0.4
requests==2.31.0
rpds-py==0.9.0
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.12.0
sentry-sdk==2.2.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sortedcontainers==2.4.0
stack-data==0.6.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
tifffile==2024.1.30
timm==1.0.3
torch==2.2.0
torchaudio==2.2.0
torchelastic==0.2.2
torchvision==0.17.0
tornado==6.4
traitlets==5.14.3
triton==2.2.0
types-dataclasses==0.6.6
typing_extensions==4.12.0
urllib3==1.26.16
wandb==0.17.0
wcwidth==0.2.13
Werkzeug==3.0.3
